--- app_main/admin.py
@@ -1,5 +1,0 @@
-from django.contrib import admin
-
-from .models import Post
-
-admin.site.register(Post)
