--- app_main/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class appmainConfig(AppConfig):
-    name = 'app_main'
