--- app_main/urls.py
@@ -1,7 +1,0 @@
-from django.urls import path
-from . import views
-
-urlpatterns = [
-    #path('', views.home, name = 'app-home'),
-    #path('about/', views.about, name='app-about'),
-]
