+++ grade_control/urls.py
@@ -15,15 +15,15 @@ Including another URLconf
 """
 from django.contrib import admin
 from django.contrib.auth import views as auth_views
-from django.urls import path
-from users import views as users_views
-from app_main import views as app_main_views
+from django.urls import path, include
+
+from system import views as users_views
 
 urlpatterns = [
     path('admin/', admin.site.urls),
     path('register/', users_views.register, name='register'),
     path('profile/', users_views.profile, name='profile'),
-    path('login/', auth_views.LoginView.as_view(template_name='users/login.html'), name='login'),
-    path('logout/', auth_views.LogoutView.as_view(template_name='users/logout.html'), name='logout'),
-    path('', app_main_views.home, name = 'home'),
+    path('login/', auth_views.LoginView.as_view(template_name='system/login.html'), name='login'),
+    path('logout/', auth_views.LogoutView.as_view(template_name='system/logout.html'), name='logout'),
+    path('', include('system.urls')),
 ]
