+++ system/templates/system/about.html
@@ -1,7 +1,7 @@
 
 <!--
 
-{% extends "app_main/base.html" %}
+{% extends "system/base.html" %}
 {% block content %}
     <h1>About Page</h1>
 {% endblock content %}
