+++ system/templates/system/home.html
@@ -1,4 +1,4 @@
-{% extends "app_main/base.html" %}
+{% extends "system/base.html" %}
 {% block content %}
     {% for grade in grades %}
         <article class="media content-section">
