+++ system/templates/system/login.html
@@ -1,4 +1,4 @@
-{% extends "app_main/base.html" %}
+{% extends "system/base.html" %}
 {% load crispy_forms_tags %}
 {% block content %}
     <div class="content-section">
