+++ system/templates/system/logout.html
@@ -1,4 +1,4 @@
-{% extends "app_main/base.html" %}
+{% extends "system/base.html" %}
 {% block content %}
     <h2>You have been logged out</h2>
     <div class="border-top pt-3">
