+++ system/templates/system/profile.html
@@ -1,4 +1,4 @@
-{% extends "app_main/base.html" %}
+{% extends "system/base.html" %}
 {% load crispy_forms_tags %}
 {% block content %}
    <h1>{{ user.username }}</h1>
