+++ system/urls.py
@@ -1,0 +1,8 @@
+from django.urls import path
+from . import views
+
+urlpatterns = [
+    path('', views.home, name='home'),
+    # path('about/', views.about, name='app-about'),
+
+]
