+++ system/views.py
@@ -1,6 +1,8 @@
 from django.shortcuts import render,  redirect
 from django.contrib import messages
 from django.contrib.auth.decorators import login_required
+
+from system.models import Grade
 from .forms import UserRegisterForm
 
 def register(request):
@@ -12,8 +14,17 @@ def register(request):
             return redirect('login')
     else:
         form = UserRegisterForm()
-    return render(request, 'users/register.html', {'form': form})
+    return render(request, 'system/register.html', {'form': form})
 
 @login_required
 def profile(request):
-    return render(request, 'users/profile.html')
+    return render(request, 'system/profile.html')
+
+
+@login_required
+def home(request):
+    context = {
+        'grades': Grade.objects.all(),
+        'title': 'gradepage'
+    }
+    return render(request, 'system/home.html', context)
